requirements.txt
handcalcs
numpy
scipy
pandas
matplotlib
forallpeople
streamlit_drawable_canvas
pillow